// src/app/pages/05-comercial/interfaces/cotizaciones.interface.ts
export interface ComercialCotizacionesCabeceraInterface {
    IdCabeceraCotizacion: number;
    FechaRegistro: Date;
    RazonSocialCliente: string;
    IdentificacionCliente: string;
    Direccioncliente: string;
    CiudadCliente: string;
    ContactoCliente: string;
    TelefonoCliente: string;
    IdentificacionAsesor: string;
    NombreAsesor: string;
    FechaPedido: Date;
    SecuencialCotizacion: number;
    TotalItems: number;
    ValorTotalCotizacion: number;
    Estado: number;
    DescripcionEstadoCabeceraCotizacion: string;
}

export interface ComercialCotizacionesDetalleInterface {
    IdDetalleCotizacion: number;
    IdCabeceraCotizacion: number;
    TipoTrabajo: string;
    TipoInventario: string;
    DescripcionTrabajo: string;
    ObservacionTrabajo: string;
    CantidadPorFabricar: number;
    UnidadMedida: string;
    UnidadesPorRollo: number;
    Rebobinado: string;
    Cono: string;
    TipoCorte: string;
    TipoCorte2: string;
    Avance: number;
    Ancho: number;
    FilasPorRollo: number;
    Troquel: string;
    GapAvance: number;
    GapAncho: number;
    Totalm2Etiqueta: number;
    Totalm2: number;
    PrecioM2Mp: number;
    PrecioTotalM2Mp: number;
    TiempoEstimadoProduccionHora: number;
    ValorHoraMaquina: number;
    ValorTotalMaquina: number;
    Color1: string;
    PorcentajeColor1: number;
    GramosColor1: number;
    GramosTotalColor1: number;
    ValorGramoColor1: number;
    ValorTotalColor1: number;
    Color2: string;
    PorcentajeColor2: number;
    GramosColor2: number;
    GramosTotalColor2: number;
    ValorGramoColor2: number;
    ValorTotalColor2: number;
    Color3: string;
    PorcentajeColor3: number;
    GramosColor3: number;
    GramosTotalColor3: number;
    ValorGramoColor3: number;
    ValorTotalColor3: number;
    Color4: string;
    PorcentajeColor4: number;
    GramosColor4: number;
    GramosTotalColor4: number;
    ValorGramoColor4: number;
    ValorTotalColor4: number;
    Color5: string;
    PorcentajeColor5: number;
    GramosColor5: number;
    GramosTotalColor5: number;
    ValorGramoColor5: number;
    ValorTotalColor5: number;
    Color6: string;
    PorcentajeColor6: number;
    GramosColor6: number;
    GramosTotalColor6: number;
    ValorGramoColor6: number;
    ValorTotalColor6: number;
    Color7: string;
    PorcentajeColor7: number;
    GramosColor7: number;
    GramosTotalColor7: number;
    ValorGramoColor7: number;
    ValorTotalColor7: number;
    Color8: string;
    PorcentajeColor8: number;
    GramosColor8: number;
    GramosTotalColor8: number;
    ValorGramoColor8: number;
    ValorTotalColor8: number;
    Color9: string;
    PorcentajeColor9: number;
    GramosColor9: number;
    GramosTotalColor9: number;
    ValorGramoColor9: number;
    ValorTotalColor9: number;
    Color10: string;
    PorcentajeColor10: number;
    GramosColor10: number;
    GramosTotalColor10: number;
    ValorGramoColor10: number;
    ValorTotalColor10: number;
    Color11: string;
    PorcentajeColor11: number;
    GramosColor11: number;
    GramosTotalColor11: number;
    ValorGramoColor11: number;
    ValorTotalColor11: number;
    Color12: string;
    PorcentajeColor12: number;
    GramosColor12: number;
    GramosTotalColor12: number;
    ValorGramoColor12: number;
    ValorTotalColor12: number;
    ValorTotalEtiqueta: number;
    ValorTotalMillar: number;
    ValorTotalItem: number;
    Estado: number;
    DescripcionDetalleCotizacion: string;
    ConfigurarOp: boolean;
    DescripcionAcabado: string;
    PrecioAcabado: number;
    CodigoProductoTerminado: string;
    EspecificacionTipoBFD: string;
    CodigoEspecificacionTipoBFD: string;
    CodigoAntiguo: string;
    FormatoTroquel: string,
    Prepicado: boolean,
    CorteSeguridad: boolean,
    Distribuidor: boolean,
}